AST/Sema: Sink isExported() queries from Sema to AST.
